Show RT numbers by default
The vanilla way of showing RT count is pretty useless,
and probably some of the worst design in the game TBH.

It took me like 100 hours to find out what it even meant,
though it might just be me being stupid.

Regardless, the number version is clearly superior, and is
much more informative than a bunch of dots.

It's also hard to read the dots if your team has more than 5 RTs.